Fixed all (- setInterval Issue) + features
All code integrated as of Tuesday

SetInterval issue from l3 has not been fixed.

Fixed loading issues and added a loading screen for the split second that the images are being loaded.

Fixed issue where part of map was erased on l6 when climbing fire escape after combining levels

Improved animations for going through tube in l2 to l11 and back

Added animation for changing lab floors in l7

Added overloading option for the drawMap function to enable it to be drawn without it calling the drawPMap function with it

Fixed Issue where music was still playing from l3 when exiting to another level