Fix++; Img--; Enemy++; Glitch--;
Fixed Issue where player would not display when hitting space on level 5 when located in the very last column.

Erased placeholder and tester images in level 4, 6 and 1.

Added a few enemies to the roof level for perspective.
**Hit detection needs to be adjusted for these and the mice**
**Ryn has changed code so that the enemies do not all look the same -- needs to replace current code for level 6**

Fixed a glitch where the a portion of the fire escape would not be drawn when walking up it and where the player would appear in several spaces on the map during and after walking up the fire escape.

Adjusted level 1 so that the level switching and animations flow with game plan better.